dddd

+++ api/services/sentinel.bridge.service.js
@@ -502,7 +502,7 @@ class SentinelBridgeService {
                 `
                 INSERT INTO sentinel_price_history (hotel_id, room_type_id, stay_date, old_price, new_price, source, created_at)
                 SELECT 
-                    t.hid::integer, t.rid, t.sdate, c.rate, t.new_price, 'SENTINEL', NOW()
+                    t.hid::integer, t.rid::integer, t.sdate, c.rate, t.new_price, 'SENTINEL', NOW()
                 FROM UNNEST($1::text[], $2::text[], $3::date[], $4::numeric[]) AS t(hid, rid, sdate, new_price)
                 JOIN sentinel_rates_calendar c 
                     ON c.hotel_id::text = t.hid
